tutorials/r4hisllm/agent_service/requirements.txt
fastapi
requests
python-dotenv
pydantic
ollama
uvicorn
langchain
langchain_core
langchain_community
langchain-ollama
opik
opentelemetry-api 
opentelemetry-sdk 
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-otlp 
opentelemetry-instrumentation-requests
opentelemetry-exporter-prometheus